Core/Scenes: moved scene functions to SceneHandler.cpp
